HS-950: Inventory - Get tags by NodeId: Fetch a list of tags based on a given NodeId
